+++ DLibSolution/test_infostat/test_sqlitestore.cpp
@@ -126,7 +126,7 @@ BOOST_AUTO_TEST_CASE(testValues) {
 	SQLiteStatHelper::DatasetType &oSet = this->m_oset;
 	//
 	size_t count = 0;
-	int nTotal = (int)(m_nbcols * m_nbrows);
+	size_t nTotal = (size_t)(m_nbcols * m_nbrows);
 	bool bRet = pMan->find_dataset_values_count(oSet, count);
 	BOOST_CHECK(bRet);
 	BOOST_CHECK(nTotal == count);
